+++ executorlib/__init__.py
@@ -12,6 +12,9 @@ Finally, the get_cache_data() function allows users to cache the content of thei
 pandas.DataFrame.
 """
 
+from typing import Optional
+
+import executorlib._version
 from executorlib.executor.base import BaseExecutor
 from executorlib.executor.flux import (
     FluxClusterExecutor,
@@ -22,12 +25,48 @@ from executorlib.executor.slurm import (
     SlurmClusterExecutor,
     SlurmJobExecutor,
 )
-from executorlib.standalone.cache import get_cache_data
 
-from . import _version
+
+def get_cache_data(cache_directory: str) -> list[dict]:
+    """
+    Collect all HDF5 files in the cache directory
+
+    Args:
+        cache_directory (str): The directory to store cache files.
+
+    Returns:
+        list[dict]: List of dictionaries each representing on of the HDF5 files in the cache directory.
+    """
+    from executorlib.standalone.hdf import get_cache_data
+
+    return get_cache_data(cache_directory=cache_directory)
+
+
+def terminate_tasks_in_cache(
+    cache_directory: str,
+    config_directory: Optional[str] = None,
+    backend: Optional[str] = None,
+):
+    """
+    Delete all jobs stored in the cache directory from the queuing system
+
+    Args:
+        cache_directory (str): The directory to store cache files.
+        config_directory (str, optional): path to the config directory.
+        backend (str, optional): name of the backend used to spawn tasks ["slurm", "flux"].
+    """
+    from executorlib.task_scheduler.file.queue_spawner import terminate_tasks_in_cache
+
+    return terminate_tasks_in_cache(
+        cache_directory=cache_directory,
+        config_directory=config_directory,
+        backend=backend,
+    )
+
 
 __all__: list[str] = [
     "get_cache_data",
+    "terminate_tasks_in_cache",
     "BaseExecutor",
     "FluxJobExecutor",
     "FluxClusterExecutor",
@@ -36,11 +75,4 @@ __all__: list[str] = [
     "SlurmClusterExecutor",
 ]
 
-try:
-    from executorlib.task_scheduler.file.queue_spawner import terminate_tasks_in_cache
-
-    __all__ += ["terminate_tasks_in_cache"]
-except ImportError:
-    pass
-
-__version__ = _version.__version__
+__version__ = executorlib._version.__version__
