+++ executorlib/task_scheduler/file/queue_spawner.py
@@ -3,9 +3,9 @@ from typing import Optional
 
 from pysqa import QueueAdapter
 
+from executorlib.standalone.hdf import dump, get_queue_id
 from executorlib.standalone.inputcheck import check_file_exists
 from executorlib.standalone.scheduler import pysqa_execute_command, terminate_with_pysqa
-from executorlib.task_scheduler.file.hdf import dump, get_queue_id
 
 
 def execute_with_pysqa(
