+++ executorlib/task_scheduler/file/subprocess_spawner.py
@@ -3,8 +3,8 @@ import subprocess
 import time
 from typing import Optional
 
+from executorlib.standalone.hdf import dump
 from executorlib.standalone.inputcheck import check_file_exists
-from executorlib.task_scheduler.file.hdf import dump
 
 
 def execute_in_subprocess(
