+++ tests/test_cache_backend_execute.py
@@ -7,7 +7,7 @@ import unittest
 try:
     from executorlib.task_scheduler.file.backend import backend_execute_task_in_file
     from executorlib.task_scheduler.file.shared import _check_task_output, FutureItem
-    from executorlib.task_scheduler.file.hdf import dump, get_runtime
+    from executorlib.standalone.hdf import dump, get_runtime
     from executorlib.standalone.serialize import serialize_funct
 
     skip_h5io_test = False
