+++ tests/test_standalone_hdf.py
@@ -4,7 +4,7 @@ import unittest
 
 
 try:
-    from executorlib.task_scheduler.file.hdf import (
+    from executorlib.standalone.hdf import (
         dump,
         load,
         get_output,
